ticketapp/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-ckeditor==6.7.2
django-debug-toolbar==5.1.0
django-js-asset==3.1.2
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
